+++ src/js/src/components/AppFooter.vue
@@ -9,6 +9,16 @@ component.
     <div class="container is-fluid">
       <small>
         <p style="float: left; color: whitesmoke; padding-bottom: 16px">
+          <b-tooltip label="phēnix documentation" type="is-light is-right">
+            <a
+              class="docs-link"
+              :href="DOCS_URL"
+              target="_blank"
+              rel="noopener"
+              aria-label="phēnix documentation">
+              <b-icon icon="question-circle" />
+            </a>
+          </b-tooltip>
           Copyright &copy; <b>2019-2026 Sandia National Laboratories</b>. All
           Rights Reserved.
         </p>
@@ -20,8 +30,12 @@ component.
 
 <script>
   import { formatVersion } from '@/utils/version.js';
+  import { DOCS_URL } from '@/utils/docs.js';
 
   export default {
+    setup() {
+      return { DOCS_URL };
+    },
     async created() {
       try {
         let resp = await fetch(this.$router.resolve({ name: 'version' }).href);
@@ -40,3 +54,15 @@ component.
     },
   };
 </script>
+
+<style scoped>
+  .docs-link {
+    color: whitesmoke;
+    margin-right: 0.5rem;
+    vertical-align: middle;
+  }
+
+  .docs-link:hover {
+    color: #ffffff;
+  }
+</style>